Refactored http module
