--- modules/http/src/test/resources/iam_json/user.json
@@ -1,33 +1,0 @@
-{
-  "identities": [
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/groups/bbp-user-one",
-      "realm": "BBP",
-      "group": "bbp-user-one",
-      "@type": "GroupRef"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/groups/bbp-svc-two",
-      "group": "bbp-svc-two",
-      "realm": "BBP",
-      "@type": "GroupRef"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/anonymous",
-      "@type": "Anonymous"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/users/f:434t3-134e-4444-aa74-bdf00f48dfce:some",
-      "realm": "BBP",
-      "sub": "f:434t3-134e-4444-aa74-bdf00f48dfce:some",
-      "@type": "UserRef"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/authenticated",
-      "realm": "BBP",
-      "@type": "AuthenticatedRef"
-    }
-  ],
-  "@type": "AuthenticatedUser",
-  "@context": "https://bbp-nexus.epfl.ch/dev/v0/contexts/bbp/core/context/v0.1.0"
-}
