--- modules/http/src/test/resources/iam_json/user_ordered.json
@@ -1,33 +1,0 @@
-{
-  "@context": "https://bbp-nexus.epfl.ch/dev/v0/contexts/bbp/core/context/v0.1.0",
-  "@type": "AuthenticatedUser",
-  "identities": [
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/groups/bbp-user-one",
-      "@type": "GroupRef",
-      "realm": "BBP",
-      "group": "bbp-user-one"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/groups/bbp-svc-two",
-      "@type": "GroupRef",
-      "realm": "BBP",
-      "group": "bbp-svc-two"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/anonymous",
-      "@type": "Anonymous"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/users/f:434t3-134e-4444-aa74-bdf00f48dfce:some",
-      "@type": "UserRef",
-      "realm": "BBP",
-      "sub": "f:434t3-134e-4444-aa74-bdf00f48dfce:some"
-    },
-    {
-      "@id": "https://nexus.example.com/v0/realms/BBP/authenticated",
-      "@type": "AuthenticatedRef",
-      "realm": "BBP"
-    }
-  ]
-}
