--- modules/http/src/test/resources/kg_json/activity_instance_att_ordered.json
@@ -1,44 +1,0 @@
-{
-  "@context": {
-    "@vocab": "https://bbp-nexus.epfl.ch/vocabs/bbp/neurosciencegraph/core/v0.1.0/",
-    "brainLocation": "https://bbp-nexus.epfl.ch/vocabs/bbp/neurosciencegraph/core/v0.1.0/brainLocation",
-    "brainRegion": "https://bbp-nexus.epfl.ch/vocabs/bbp/neurosciencegraph/core/v0.1.0/brainRegion",
-    "description": "http://schema.org/description",
-    "label": "http://www.w3.org/2000/01/rdf-schema#label",
-    "name": "http://schema.org/name",
-    "nsg": "https://bbp-nexus.epfl.ch/vocabs/bbp/neurosciencegraph/core/v0.1.0/",
-    "providerId": "https://bbp-nexus.epfl.ch/vocabs/bbp/neurosciencegraph/core/v0.1.0/providerId",
-    "wasDerivedFrom": "http://www.w3.org/ns/prov#wasDerivedFrom"
-  },
-  "@id": "https://bbp-nexus.epfl.ch/dev/v0/data/bbp/experiment/patchedcell/v0.1.0/a2abf855-53af-4d4a-a601-04d6029591dc",
-  "@type": [
-    "prov:Entity",
-    "nsg:PatchedCell"
-  ],
-  "brainLocation": {
-    "brainRegion": {
-      "@id": "http://www.FIXME.org/whs_sd_ontology/0000014",
-      "label": "Somatosensory cortex hindlimb layer IV"
-    }
-  },
-  "description": "original layer information: L4Bott; cellid: 2707",
-  "name": "C010600A2-MT-C1",
-  "providerId": "2707",
-  "nxv:rev": 4,
-  "nxv:originalFileName": "Test.docx",
-  "nxv:contentType": "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-  "nxv:size": {
-    "nxv:unit": "byte",
-    "nxv:value": 22506
-  },
-  "nxv:digest": {
-    "nxv:alg": "SHA-256",
-    "nxv:value": "84d9c6ff32560c3df9a0f1da36708550a957f012c13e03d3aaa0634050a5f531"
-  },
-  "nxv:deprecated": false,
-  "links": {
-    "@context": "https://bbp-nexus.epfl.ch/dev/v0/contexts/bbp/core/links/v0.1.0",
-    "self": "https://bbp-nexus.epfl.ch/dev/v0/data/bbp/experiment/patchedcell/v0.1.0/a2abf855-53af-4d4a-a601-04d6029591dc",
-    "schema": "https://bbp-nexus.epfl.ch/dev/v0/schemas/bbp/experiment/patchedcell/v0.1.0"
-  }
-}
